Remove downstream dependency on true
Trading the location of a single line of code may not seem like much, however `true` will, for reasons unknown, not build when the file path includes a space due to some upstream issue with ExtUtils::Depends.

Would have patched upstream, but in all honesty removing a dependency from the river seems like a bigger win.